impl more map functions

+++ sokogen/consts.c
@@ -1,9 +1,9 @@
 #include "consts.h"
 
-const int texWidth = 32;
-const int texHeight = 32;
+const int8_t texWidth = 32;
+const int8_t texHeight = 32;
 
-const short DirectionOff[4][2] = {
+const int8_t DirectionOff[4][2] = {
 	{0, -1},
 	{0, 1},
 	{-1, 0},
